feat: update spec to 0.1.10
Problem:
New version of API spec is available for Text2Motion with support for
new experimental model.

Solution:
Update the API spec to the latest version.

Risk:
This change added new API, which is a non-breaking change for existing
clients.

Testing Done:
Tested generating python-client using the new spec and successfully use
it to make server request to the existing generate API and also the new
lab generate API.